requirements.txt
Django==4.2
djangorestframework==3.14.0
flake8==6.0.0
flake8-bugbear==21.4.3
autopep8==1.5.7
